standardize file names
