+++ optimum/bettertransformer/models/bart/__init__.py
@@ -11,4 +11,4 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from .modeling_bart_fast import BartLayerBetterTransformer
+from .modeling_bart_bettertransformer import BartLayerBetterTransformer
