+++ optimum/bettertransformer/models/bert/__init__.py
@@ -11,4 +11,4 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from .modeling_bert_fast import BertLayerBetterTransformer
+from .modeling_bert_bettertransformer import BertLayerBetterTransformer
