+++ optimum/bettertransformer/models/distilbert/__init__.py
@@ -11,4 +11,4 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-from .modeling_distilbert_fast import DistilBertLayerBetterTransformer
+from .modeling_distilbert_bettertransformer import DistilBertLayerBetterTransformer
